requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.1.0
billiard==4.2.0
black==24.3.0
boto3==1.34.77
botocore==1.34.77
cachetools==5.3.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.4
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.13
django-allauth==0.61.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.3.0
django-filter==23.5
django-redis==5.4.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==7.0.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==22.7.0
flake8==7.0.0
flower==2.0.1
gevent==24.2.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gspread==6.0.2
gunicorn==21.2.0
h11==0.14.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.5.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.0
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0

PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.1
selenium==4.18.1
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.4.4
StrEnum==0.4.15
text-unidecode==1.3
tornado==6.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
wsproto==1.2.0
zope.event==5.0
zope.interface==6.2
